# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "description": "API documentation for TaskAPI",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root Info Endpoint",
        "description": "Returns basic API metadata",
        "responses": {
          "200": {
            "description": "API details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string", "example": "Task API" },
                    "version": { "type": "string", "example": "1.0" },
                    "enpoints": {
                      "type": "array",
                      "items": { "type": "string" },
                      "example": ["/tasks"]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns operational health status",
        "responses": {
          "200": {
            "description": "Server health status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Get task Stats",
        "description": "Returns stats about tasks",
        "responses": {
          "200": {
            "description": "Task statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer", "example": 10 },
                    "done": { "type": "integer", "example": 5 },
                    "open": { "type": "integer", "example": 5 }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {

      "get": {
        "summary": "Get All Tasks or Filter by Done Status",
        "description": "Retrieves list of all tasks",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Filter tasks by done title",
            "schema": { "type": "string" }
          },
          {
            "name": "done",
            "in": "query",
            "description": "Filter tasks by done status",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Task" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Task",
        "description": "Creates a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "Learn Swagger" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "newtask": { "$ref": "#/components/schemas/Task" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid title",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Bad Request : Please define a title" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Internal Server Error" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get Task By ID",
        "description": "Retrieves a single task by its numeric ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric ID of task",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task object found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task": { "$ref": "#/components/schemas/Task" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Error Task 1 not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Task By ID",
        "description": "Updates an existing task's title or status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric ID of task",
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Updated Task Title" },
                  "done": { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updatedTask": { "$ref": "#/components/schemas/Task" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid title input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Bad Request: Invalid Body" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Not found" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Task By ID",
        "description": "Deletes a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric ID of task",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully (No content)"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Not found" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 0 },
          "title": { "type": "string", "example": "Sample Task" },
          "done": { "type": "boolean", "example": false }
        }
      }
    }
  }
}